Road Damage Detection using ResNet50/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
scipy>=1.5.0
scikit-learn>=0.24.0
scikit-image>=0.18.0
matplotlib>=3.3.0
seaborn>=0.11.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=8.0.0
tqdm>=4.60.0
pyyaml>=5.4.0
statsmodels>=0.12.0
